New notebook will have comparison plots.
For now it runs the MC control with the quadrature method. The output here should
be pretty similar to that in scribble_quadrature, but somehow the first episode
already has a different length, so probably some randomness is not quite controlled.

This version shall also act as a "gold standard" for some coming optimizations.